+++ .gitignore
@@ -1,6 +1,5 @@
 .gradle
-/local.properties
-/.idea/workspace.xml
+local.properties
 .DS_Store
 .idea
 build
