+++ Fevi/src/main/java/com/app/fevir/interfaces/web/presenter/OpenPresenter.java
@@ -2,9 +2,11 @@ package com.app.fevir.interfaces.web.presenter;
 
 import com.app.fevir.movie.list.domain.Card;
 
+import rx.functions.Action1;
+
 public interface OpenPresenter {
 
-    void onInit(String lastPath);
+    void onInit(String lastPath, Action1<? super Card> subscribe);
 
     interface View {
