+++ Fevi/src/main/java/com/app/fevir/interfaces/web/presenter/OpenPresenterImpl.java
@@ -1,6 +1,7 @@
 package com.app.fevir.interfaces.web.presenter;
 
 import com.app.fevir.interfaces.web.model.OpenModel;
+import com.app.fevir.movie.list.domain.Card;
 import com.app.fevir.network.api.Cards;
 
 import javax.inject.Inject;
@@ -8,6 +9,7 @@ import javax.inject.Inject;
 import retrofit.Retrofit;
 import rx.Observable;
 import rx.android.schedulers.AndroidSchedulers;
+import rx.functions.Action1;
 
 public class OpenPresenterImpl implements OpenPresenter {
     private final View view;
@@ -22,7 +24,7 @@ public class OpenPresenterImpl implements OpenPresenter {
     }
 
     @Override
-    public void onInit(String lastPath) {
+    public void onInit(String lastPath, Action1<? super Card> subscribe) {
         if (openModel.containHashUrl(lastPath)) {
 
             retrofit.create(Cards.class)
@@ -30,7 +32,7 @@ public class OpenPresenterImpl implements OpenPresenter {
                     .flatMap(cardInfo -> Observable.from(cardInfo.getContent()))
                     .first()
                     .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe(view::showDetail, Throwable::printStackTrace);
+                    .subscribe(subscribe, Throwable::printStackTrace);
 
         } else {
             view.showHome();
