+++ Fevi/src/main/java/com/app/fevir/movie/detail/presenter/MovieDetailPresenter.java
@@ -14,7 +14,7 @@ public interface MovieDetailPresenter {
 
     interface View {
 
-        void sendFacebookLink(String shareMsg);
+        void sendWebLink(String shareMsg);
 
         void sendWeb(String url);
     }
