+++ Fevi/src/main/res/layout/activity_movie.xml
@@ -19,7 +19,7 @@
 
         <VideoView
             android:id="@+id/fa_picture"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="200dp"
             android:layout_centerInParent="true"
             android:layout_centerVertical="true" />
@@ -46,7 +46,7 @@
     <RelativeLayout
         android:id="@+id/profile_layout"
         android:layout_width="match_parent"
-        android:layout_height="?actionBarSize"
+        android:layout_height="wrap_content"
         android:layout_below="@+id/video_layout"
         android:background="#fff">
 
@@ -54,14 +54,13 @@
         <ImageView
             android:id="@+id/fa_profile"
             android:layout_width="50dp"
-            android:layout_height="match_parent"
+            android:layout_height="50dp"
             android:layout_margin="5dp" />
 
         <TextView
             android:id="@+id/fa_name"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_above="@+id/fa_time"
             android:layout_alignParentTop="true"
             android:layout_marginRight="5dp"
             android:layout_marginTop="10dp"
@@ -73,39 +72,35 @@
             android:id="@+id/fa_time"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_marginBottom="10dp"
+            android:layout_alignBottom="@+id/fa_profile"
             android:layout_marginRight="5dp"
             android:layout_toRightOf="@+id/fa_profile"
             android:textSize="10sp" />
 
+        <ScrollView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_below="@+id/fa_profile"
+            android:fillViewport="true">
+
+
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:background="#fff"
+                android:padding="10dp">
+
+                <TextView
+                    android:id="@+id/fa_description"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:ellipsize="end"
+                    android:text="동영상 소개내용들..."
+                    android:textColor="#4e4e4e" />
+
+            </RelativeLayout>
+        </ScrollView>
 
     </RelativeLayout>
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/vg_description"
-        android:layout_below="@+id/profile_layout"
-        android:fillViewport="true">
-
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="#fff"
-            android:padding="10dp">
-
-            <TextView
-                android:id="@+id/fa_description"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:ellipsize="end"
-                android:text="동영상 소개내용들..."
-                android:textColor="#4e4e4e" />
-
-        </RelativeLayout>
-    </ScrollView>
-
 
 </RelativeLayout>
