+++ Fevi/src/main/res/values/styles.xml
@@ -11,6 +11,7 @@
 
         <item name="toolbarStyle">@style/VikiCast.Default.ActionBar</item>
         <item name="colorControlNormal">#FFFFFFFF</item>
+
         <item name="android:textColorSecondary">#FFFFFFFF</item>
         <item name="android:textColorPrimary">#FFFFFFFF</item>
     </style>
@@ -19,6 +20,7 @@
     <style name="VikiCast.Default.ActionBar" parent="@style/Widget.AppCompat.Toolbar">
         <item name="android:background">?attr/colorPrimary</item>
         <item name="android:layout_height">?attr/actionBarSize</item>
+        <item name="popupTheme">@style/ThemeOverlay.AppCompat.Light</item>
     </style>
 
     <style name="VikiCast.Toolbar.Title" parent="@style/TextAppearance.Widget.AppCompat.Toolbar.Title">
